upload
This repository contains my resume, detailing my educational background, skills, projects, and experiences in software development, data science, and robotics. It showcases my journey as a programmer and highlights my commitment to continuous learning and professional growth.

Feel free to explore the document and connect with me for potential collaborations or opportunities!